fix: account for when no good color can be found
